requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==24.2
PyJWT==2.10.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
